#saved: created components directory and moved drink list to it
